FIX: double promotion in demo printf

+++ demo/adc_noise_measurement/src/main.c
@@ -186,7 +186,7 @@ int main(void)
     uint32_t duration = end_time - start_time;
     printf("Total Time: %u ms\n", duration);
     if(duration > 0) {
-        printf("Frequency: %.2f Hz\n", (float)sample_count * 1000.0f / (float)duration);
+        printf("Frequency: %.2f Hz\n", (double)sample_count * 1000.0 / (double)duration);
     }
